Fake deserializing >2**31 issue comment IDs
We have recently hit larger IDs in production, this tries to temporarily
work around that by making use of the negative ID space, hopefully
buying us a good chunk of time to make the larger change to switch to
2**64 IDs. Since the conversion is bidirectional, if not numeric value
preserving, it should hopefully be OK.

+++ src/github/client.rs
@@ -208,7 +208,7 @@ impl Client {
     ) -> DashResult<CommentFromJson> {
         let url = format!(
             "{}/repos/{}/issues/comments/{}",
-            BASE_URL, repo, comment_num
+            BASE_URL, repo, comment_num as u32
         );
         let payload = serde_json::to_string(&btreemap!("body" => text))?;
         Ok(self.patch(&url, &payload)?.error_for_status()?.json()?)
